rearranged components, created pages folder
